refactor(admin): extract DemoTab into its own file
